Re-generate wasi-types bindings
The last commit refactored the bindings generation code.
This commit now actually contains re-generated bindings built with the new
generator.
