+++ lib/wasi-types/src/wasi/mod.rs
@@ -1,7 +1,7 @@
-pub(crate) mod extra;
-pub(crate) mod extra_manual;
-pub use extra::*;
-pub use extra_manual::*;
+pub(crate) mod bindings;
+pub(crate) mod bindings_manual;
+pub use self::bindings::*;
+pub use bindings_manual::*;
 
 mod wasix_manual;
 pub use wasix_manual::*;
